Resolve all TODOs
