+++ CHANGELOG.md
@@ -11,7 +11,8 @@ CHANGELOG
       - `worksheet.add_border "A1", {style: :thin}`
       - `worksheet.add_border "A1:B2", {style: :thin}`
       - `worksheet.add_border ["A1", "B2:C7", "D8:E9"], {style: :thin}`
-    - Add `BorderCreator` - TODO: Do we mention this at all?
+    - Add `Axlsx::BorderCreator` the class used under the hood for `worksheet.add_border`
+    - Allow specifying `:all` in `border: {edges: :all}` which is a shortcut for `border: {edges: [:left, :right, :top, :bottom]}`
 
 - **Unreleased**
   - [PR #155](https://github.com/caxlsx/caxlsx/pull/155) - Add `hideDropDown` alias for `showDropDown` setting, as the latter is confusing to use (because its logic seems inverted).
